Rename delete button and extract getBooks into Books component
