+++ src/components/books.js
@@ -1,8 +1,13 @@
-import React from 'react';
+import React, { useEffect } from 'react';
 import { useSelector } from 'react-redux';
 import OneBook from './oneBook';
+import store from '../redux/configureStore';
+import { getBooks } from '../redux/books/books';
 
 const Books = () => {
+  useEffect(() => {
+    store.dispatch(getBooks());
+  }, []);
   const books = useSelector((state) => state.books);
 
   return (
