+++ src/index.js
@@ -3,10 +3,6 @@ import ReactDOM from 'react-dom';
 import { BrowserRouter } from 'react-router-dom';
 import './index.css';
 import App from './App';
-import store from './redux/configureStore';
-import { getBooks } from './redux/books/books';
-
-store.dispatch(getBooks());
 
 ReactDOM.render(
   <React.StrictMode>
